backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
rembg==2.0.56
pillow==10.2.0
numpy==1.26.4
onnxruntime==1.17.1
google-genai

